net: Fix convert_str_to_ssaddr
The function returns EINVAL if the string contains both an address and a port,
but also specifies a non-zero default port. This behavior is incorrect—
if the string already includes a port, the default port should be ignored.
The default port should only serve as a fallback when no port is specified
in the string.

And add header guard in net.h

Fixes e6bd4242be57 ("dnslookup: Add a new parameter default_port")
Signed-off-by: Ahmad Gani <reyuki@gnuweeb.org>
Link: https://lore.gnuweeb.org/gwml/20250816103807.27781-1-reyuki@gnuweeb.org
Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>